move NewTransaction to package transaction
