+++ app/src/main/java/com/ahmadrosid/dompetku/main/MainActivity.java
@@ -14,7 +14,7 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.ahmadrosid.dompetku.DetailTransactionActivity;
-import com.ahmadrosid.dompetku.NewTransaction;
+import com.ahmadrosid.dompetku.transaction.NewTransaction;
 import com.ahmadrosid.dompetku.R;
 import com.ahmadrosid.dompetku.data.Ballance;
 import com.ahmadrosid.dompetku.data.Transactions;
