+++ app/src/main/java/com/ahmadrosid/dompetku/transaction/NewTransaction.java
@@ -1,4 +1,4 @@
-package com.ahmadrosid.dompetku;
+package com.ahmadrosid.dompetku.transaction;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -11,6 +11,8 @@ import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.Spinner;
 
+import com.ahmadrosid.dompetku.R;
+import com.ahmadrosid.dompetku.StateBottomeSet;
 import com.ahmadrosid.dompetku.data.Transactions;
 
 import io.realm.Realm;
@@ -33,11 +35,14 @@ public class NewTransaction extends BottomSheetDialogFragment implements View.On
 
     private Realm realm;
 
-    @Nullable @Override public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         return inflater.inflate(R.layout.new_transaction_bottomset, container, false);
     }
 
-    @Override public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
+    @Override
+    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
         item_name = (EditText) view.findViewById(R.id.item_name);
@@ -53,7 +58,8 @@ public class NewTransaction extends BottomSheetDialogFragment implements View.On
         transaction.setAdapter(adapter);
     }
 
-    @Override public void onClick(View view) {
+    @Override
+    public void onClick(View view) {
         switch (view.getId()) {
             case R.id.img_close:
                 dismiss();
